Add routers for routing api calls
We are going to use routers to route API calls. The outline is as
follows. We make app use the baseUrl (/observatory/api) and from there
we call MainRouter. MainRouter will call another router depending on
the next url path. There will be at least 3 such routers. One for
products, one for shops and one for prices, each called for the
corresponding path.