Complete Session Manager
Sessions are managed through this class. When a new session is needed,
NewSession() must be called with the user as an argument. This will
create a session_info entry in the session_list of the SessionManager. A
check can be performed by calling CheckSession() with the token as an
argument. This returns the session_info on success and 'none' on
failure (perhaps should return something else on failure?). The session
can be ended by calling EndSession() with the token as an argument. This
returns true on success (whether the session existed or not).